package store split to logic and IStore interface
let store support database, cache or file storage interface.
