+++ core/store/store.go
@@ -1,15 +1,24 @@
 package store
 
-import(
-	. "DNA/core/ledger"
-	. "DNA/core/store/LevelDBStore"
-)
-
-func NewLedgerStore() ILedgerStore {
-	// TODO: read config file decide which db to use.
-	ldbs,_ := NewLevelDBStore("Chain")
-
-	return ldbs
+type IIterator interface {
+	Next() bool
+	Prev() bool
+	First() bool
+	Last() bool
+	Seek(key []byte) bool
+	Key() []byte
+	Value() []byte
+	Release()
 }
 
-
+type IStore interface {
+	Put(key []byte, value []byte) error
+	Get(key []byte) ([]byte, error)
+	Delete(key []byte) error
+	NewBatch() error
+	BatchPut(key []byte, value []byte) error
+	BatchDelete(key []byte) error
+	BatchCommit() error
+	Close() error
+	NewIterator(prefix []byte) IIterator
+}
